amf, swf二进制格式

+++ sort/bubble/bubble.go
@@ -7,10 +7,10 @@ func Sort(sample []int) {
 }
 
 func theory(sample []int) {
-	n := uint(len(sample))
+	n := len(sample)
 
-	for i := uint(1); i < n; i++ {
-		for j := uint(1); j < n; j++ {
+	for i := 1; i < n; i++ {
+		for j := 1; j < n; j++ {
 			if sample[j-1] > sample[j] {
 				sample[j-1], sample[j] = sample[j], sample[j-1]
 			}
@@ -19,12 +19,12 @@ func theory(sample []int) {
 }
 
 func optimize(sample []int) {
-	n := uint(len(sample))
+	n := len(sample)
 
-	for sorted, i := false, uint(1); !sorted && i < n; i++ {
+	for sorted, i := false, 1; !sorted && i < n; i++ {
 		sorted = true
 
-		for j := uint(1); j < n-i+1; j++ {
+		for j := 1; j < n-i+1; j++ {
 			if sample[j-1] > sample[j] {
 				sorted = false
 				sample[j-1], sample[j] = sample[j], sample[j-1]
